Added a 3rd Mono to addSale method to check the quantity of each product stock; Added a product gateway to comunicate with the Products API;

+++ src/main/java/com/letscode/ecsalesapi/domain/SaleCartRequest.java
@@ -13,6 +13,6 @@ public class SaleCartRequest {
     private Integer userId;
     private Boolean isActiveStatus;
     private BigDecimal totalPrice;
-    private Map<String, BigDecimal> products;
+    private Map<String, Long> products;
 
 }
